Module type Mirage_nat.TABLE

type t

module TCP : SUBTABLE with type t := t and type transport_channel = port * port

A TCP channel is identified by the source and destination ports.

module UDP : SUBTABLE with type t := t and type transport_channel = port * port

A UDP channel is identified by the source and destination ports.

module ICMP : SUBTABLE with type t := t and type transport_channel = Cstruct.uint16

An ICMP query is identified by the ICMP ID.

val reset : t -> unit Lwt.t

Remove all entries from the table.