Parameter Make.1-Table

type t

module TCP : Mirage_nat.SUBTABLE with type t := t and type transport_channel = Mirage_nat.port * Mirage_nat.port

A TCP channel is identified by the source and destination ports.

module UDP : Mirage_nat.SUBTABLE with type t := t and type transport_channel = Mirage_nat.port * Mirage_nat.port

A UDP channel is identified by the source and destination ports.

module ICMP : Mirage_nat.SUBTABLE with type t := t and type transport_channel = Cstruct.uint16

An ICMP query is identified by the ICMP ID.

val reset : t -> unit Lwt.t

Remove all entries from the table.

val remove_connections : t -> Ipaddr.V4.t -> Mirage_nat.ports

Remove all connections from and to the given IP address.